refactor(InputParser): 기존 문자열 파싱로직
문자열 파싱 로직 길어서 분리 (1단계 피드백)
